+++ pythonFiles/getFiltered_zero.py
@@ -35,16 +35,16 @@ yh = np.array(ds.variables['yh'])
 timeVal = np.array(ds.variables['Time'])
 timeUnits = ds.variables['Time'].units
 
-U = np.array(ds.variables['u'][:, :, :], dtype=float)
-V = np.array(ds.variables['v'][:, :, :], dtype=float)
-h = np.array(ds.variables['h'][:, :, :], dtype=float)
-P = np.array(ds.variables['e'][:, :, :], dtype=float) * 9.81 ## constant rho is omitted
+U = np.array(ds.variables['u'][:, :, :], dtype=np.float64)
+V = np.array(ds.variables['v'][:, :, :], dtype=np.float64)
+h = np.array(ds.variables['h'][:, :, :], dtype=np.float64)
+P = np.array(ds.variables['e'][:, :, :], dtype=np.float64) * 9.81 ## constant rho is omitted
 
 timeLen, Ylen, Xlen = np.shape(U)
 print('shape of the array U ', timeLen, Ylen, Xlen)
 sys.stdout.flush()
 
-farr = np.ones((Ylen, Xlen), dtype=float)
+farr = np.ones((Ylen, Xlen), dtype=np.float64)
 f0 = 6.49e-05
 beta = 2.0E-11
 
@@ -91,17 +91,17 @@ writeDS.createDimension('Time', None)
 writeDS.createDimension('xh', 240)
 writeDS.createDimension('yh', 320)
 
-wcdf_Xh = writeDS.createVariable('xh', np.float32, ('xh'))
+wcdf_Xh = writeDS.createVariable('xh', np.float64, ('xh'))
 wcdf_Xh.long_name = 'h point nominal longitude'
 wcdf_Xh.units = 'kilometers'
 wcdf_Xh[:] = xh[:]
 
-wcdf_Yh = writeDS.createVariable('yh', np.float32, ('yh'))
+wcdf_Yh = writeDS.createVariable('yh', np.float64, ('yh'))
 wcdf_Yh.long_name = 'h point nominal latitude'
 wcdf_Yh.units = 'kilometers'
 wcdf_Yh[:] = yh[:]
 
-wcdf_Time = writeDS.createVariable('Time', np.float32, ('Time'))
+wcdf_Time = writeDS.createVariable('Time', np.float64, ('Time'))
 wcdf_Time.long_name = "Time"
 wcdf_Time.units = timeUnits
 wcdf_Time.cartesian_axis = "T"
@@ -110,103 +110,103 @@ wcdf_Time.calendar = "JULIAN"
 wcdf_Time[:] = timeVal
 
 wcdf_U_bar = writeDS.createVariable(
-    'u_bar', np.float32, ('Time', 'yh', 'xh'))
+    'u_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_U_bar.long_name = "u_bar"
 wcdf_U_bar.units = "m s^-1"
 wcdf_U_bar[:, :, :] = U_bar[:, :, :]
 
 wcdf_V_bar = writeDS.createVariable(
-    'v_bar', np.float32, ('Time', 'yh', 'xh'))
+    'v_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_V_bar.long_name = "v_bar"
 wcdf_V_bar.units = "m s^-1"
 wcdf_V_bar[:, :, :] = V_bar[:, :, :]
 
 wcdf_hfU_bar = writeDS.createVariable(
-    'hfu_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hfu_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hfU_bar.long_name = "hfu_bar"
 wcdf_hfU_bar.units = "m^2 s^-2"
 wcdf_hfU_bar[:, :, :] = hfU_bar[:, :, :]
 
 wcdf_hfV_bar = writeDS.createVariable(
-    'hfv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hfv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hfV_bar.long_name = "hfv_bar"
 wcdf_hfV_bar.units = "m^2 s^-2"
 wcdf_hfV_bar[:, :, :] = hfV_bar[:, :, :]
 
 wcdf_h_bar = writeDS.createVariable(
-    'h_bar', np.float32, ('Time', 'yh', 'xh'))
+    'h_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_h_bar.long_name = "h_bar"
 wcdf_h_bar.units = "m"
 wcdf_h_bar[:, :, :] = h_bar[:, :, :]
 
 wcdf_hU_bar = writeDS.createVariable(
-    'hu_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hu_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hU_bar.long_name = "hu_bar"
 wcdf_hU_bar.units = "m^2 s^-1"
 wcdf_hU_bar[:, :, :] = hU_bar[:, :, :]
 
 wcdf_hV_bar = writeDS.createVariable(
-    'hv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hV_bar.long_name = "hv_bar"
 wcdf_hV_bar.units = "m^2 s^-1"
 wcdf_hV_bar[:, :, :] = hV_bar[:, :, :]
 
 wcdf_hUU_bar = writeDS.createVariable(
-    'huu_bar', np.float32, ('Time', 'yh', 'xh'))
+    'huu_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hUU_bar.long_name = "huu_bar"
 wcdf_hUU_bar.units = "m^3 s^-2"
 wcdf_hUU_bar[:, :, :] = hUU_bar[:, :, :]
 
 wcdf_hVV_bar = writeDS.createVariable(
-    'hvv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hvv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hVV_bar.long_name = "hvv_bar"
 wcdf_hVV_bar.units = "m^3 s^-2"
 wcdf_hVV_bar[:, :, :] = hVV_bar[:, :, :]
 
 wcdf_hUV_bar = writeDS.createVariable(
-    'huv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'huv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hUV_bar.long_name = "huv_bar"
 wcdf_hUV_bar.units = "m^3 s^-2"
 wcdf_hUV_bar[:, :, :] = hUV_bar[:, :, :]
 
 wcdf_UU_bar = writeDS.createVariable(
-    'uu_bar', np.float32, ('Time', 'yh', 'xh'))
+    'uu_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_UU_bar.long_name = "uu_bar"
 wcdf_UU_bar.units = "m^2 s^-2"
 wcdf_UU_bar[:, :, :] = UU_bar[:, :, :]
 
 wcdf_VV_bar = writeDS.createVariable(
-    'vv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'vv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_VV_bar.long_name = "vv_bar"
 wcdf_VV_bar.units = "m^2 s^-2"
 wcdf_VV_bar[:, :, :] = VV_bar[:, :, :]
 
 wcdf_UV_bar = writeDS.createVariable(
-    'uv_bar', np.float32, ('Time', 'yh', 'xh'))
+    'uv_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_UV_bar.long_name = "uv_bar"
 wcdf_UV_bar.units = "m^2 s^-2"
 wcdf_UV_bar[:, :, :] = UV_bar[:, :, :]
 
 wcdf_P_bar = writeDS.createVariable(
-    'p_bar', np.float32, ('Time', 'yh', 'xh'))
+    'p_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_P_bar.long_name = "p_bar"
 wcdf_P_bar.units = "m^2 s^-2"
 wcdf_P_bar[:, :, :] = P_bar[:, :, :]
 
 wcdf_Pdx_h_bar = writeDS.createVariable(
-    'pdx_h_bar', np.float32, ('Time', 'yh', 'xh'))
+    'pdx_h_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_Pdx_h_bar.long_name = "pdx_bar"
 wcdf_Pdx_h_bar.units = "m^2 s^-2"
 wcdf_Pdx_h_bar[:, :, :] = Pdx_h_bar[:, :, :]
 
 wcdf_Pdy_h_bar = writeDS.createVariable(
-    'pdy_h_bar', np.float32, ('Time', 'yh', 'xh'))
+    'pdy_h_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_Pdy_h_bar.long_name = "pdy_bar"
 wcdf_Pdy_h_bar.units = "m^2 s^-2"
 wcdf_Pdy_h_bar[:, :, :] = Pdy_h_bar[:, :, :]
 
 wcdf_hP_bar = writeDS.createVariable(
-    'hp_bar', np.float32, ('Time', 'yh', 'xh'))
+    'hp_bar', np.float64, ('Time', 'yh', 'xh'))
 wcdf_hP_bar.long_name = "hp_bar"
 wcdf_hP_bar.units = "m^3 s^-2"
 wcdf_hP_bar[:, :, :] = hP_bar[:, :, :]
